Login update
This lets you use it as a website, instead of random file paths in a specific computer so when we get it running online it wont crash. Homepage, contactus, Login, and buyerHomepage and cancel order work. check those for the new style in which we have to use pathing. its href url for (). this needs to be done for all of our code.