Add items
You have a VERY BASIC ADD ITEMS TO CART SYSTEM. cant check the cart(except through the database), it doesn't change the quantity available for the item in question, and it doesnt check if the items are already in the cart. that will be done on like. . . friday or something.